added targets and changed changed to unsigned int

+++ common/barrier.hpp
@@ -23,9 +23,9 @@ class barrier {
     pthread_barrier_t b_{};
 
    public:
-    explicit barrier(int num_threads) {
+    explicit barrier(unsigned int num_threads) {
         assert(num_threads > 0);
-        if (int rc = pthread_barrier_init(&b_, nullptr, static_cast<unsigned int>(num_threads)); rc != 0) {
+        if (int rc = pthread_barrier_init(&b_, nullptr, num_threads); rc != 0) {
             throw std::system_error{rc, std::system_category(), "Failed to create barrier: "};
         }
     }
